Move files to the place where they're used
