+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/FiltersModal/BottomBar.tsx
