+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/FiltersModal/TopBar.tsx
