+++ front/app/components/IdeaCards/IdeasWithFiltersSidebar/FiltersModal/messages.ts
@@ -13,4 +13,13 @@ export default defineMessages({
     id: 'app.components.FiltersModal.a11y_closeFilterPanel',
     defaultMessage: 'Close filter panel',
   },
+  showResults: {
+    id: 'app.containers.IdeaCards.showResults',
+    defaultMessage: 'Show results',
+  },
+  showXResults: {
+    id: 'app.containers.IdeaCards.showXResults',
+    defaultMessage:
+      'Show {ideasCount, plural, no {# results} one {# result} other {# results}}',
+  },
 });
